Configured 'StateRepository' bean in TogglzConfiguration class. This is required for persisting the state of the toggles (in db), otherwise changes to the toggles' state will be lost as soon as the server is restarted.
Note the use of CachingStateRepository, which will act as a cache for persistentStateRepository. It will cache the results for 'cachingStateRepositoryTtl' milliseconds. If you omit the timeout, lookups will be cached until setFeatureState() is called for the specific feature. Using the cache without a timeout only makes sense if the repository state is never modified directly (by modifying the database table for example).