Used FluentLenium, which provides a Java fluent interface to Selenium, and brings some magic to avoid common issues faced by Selenium users.
Having this library also removes the need of Spring's autowiring, as the PageObject classes are not autowired any more, but they are annotated with '@Page' (which will inject the constructed PageObject wherever required).

More importantly, there is no dirty context logic in 'SeleniumTestExecutionListener' any more, which means the tests will execute faster as reloading of context is not required any more (reloading was required with the earlier approach due to the injection of web driver in Page Objects).

NOTE: AbstractSeleniumTest is extended from FluentTest, where as BasePage is extended from FluentPage.